✨ Refactor ISCN data for payload generation
